refactor(server): move document engine into package
