+++ packages/doc-server/src/document-engine-shared.ts
@@ -1,9 +1,9 @@
 export {
   executeDocumentOperation,
   executeDocumentOperationAsync,
-} from '../../../server/document-engine.js';
+} from './document-engine.js';
 
 export type {
   AsyncDocumentMutationContext,
   EngineExecutionResult,
-} from '../../../server/document-engine.js';
+} from './document-engine.js';
